remove unused logging; wip: fix uniqueness issues
we are having some issues with our database queries causing
trajectory chunks to be run more than once. I think this is fixed
but I need to go through with a fine-tooth comb to make sure

committing now since we seem to produce correct trajectories now.
previously we had trajectories that would repeat a chunk
periodically

I think there may still be some non-uniquness sneaking in,
but not sure